fix mp3_play_file_in_folder

+++ DFPlayer_Mini_Mp3/DFPlayer_Mini_Mp3.h
@@ -98,6 +98,9 @@ void mp3_set_debug_serial (HardwareSerial &theSerial);
 //
 //void mp3_send_cmd (uint8_t cmd);
 
+// Send byte1 (0-15) as first byte of high and byte234 (0 - 4095) splitted in high byte 2 and low
+//void mp3_send_cmd (uint8_t cmd, uint8_t byte1, uint32_t byte234)
+
 // Reset buffer
 //void reset_recv_buf ();
 
@@ -219,4 +222,4 @@ void mp3_get_folder_sum (uint16_t folder);
 int mp3_wait_folder_sum ();
 
 // Play mp3 file in selected folder
-void mp3_play_file_in_folder (uint16_t folder, uint16_t num);
+void mp3_play_file_in_folder (uint8_t folder, uint32_t num);
